backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.12.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
groq>=0.9.0
# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic>=1.13.0
# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
